agents/supervisor/requirements.txt
google-adk[a2a]
fastapi>=0.104.0
uvicorn[standard]>=0.24.0